+++ src/Installers/PaymentMethodsInstaller.php
@@ -245,7 +245,7 @@ class PaymentMethodsInstaller implements InstallerInterface
             $this->setBuckarooPaymentSettingsValue($paymentMethod->getBuckarooKey(), $paymentMethod->getName(), 'Label');
         }
 
-        foreach ([['pendingPaymentStatus'=>'open'],['paymentSuccesStatus'=>'paid'],['paymentFailedStatus'=>'cancelled'],['orderStatus'=>'open']] as $key => $value) {
+        foreach ([['pendingPaymentStatus'=>'open'],['paymentSuccesStatus'=>'paid'],['paymentFailedStatus'=>'cancelled'],['orderStatus'=>'open'],['BillinkBusiness'=>'B2C']] as $key => $value) {
             $this->setBuckarooPaymentSettingsValue($key, $value);
         }
     }