+++ CHANGELOG.md
@@ -54,3 +54,4 @@ BP-206 Add payment method KlarnaKP
 
 # 1.1.2
 BP-584 Multiple issues after code review
+BP-583 Add new bank to iDEAL payment method